testing bsd: install iked.conf when present

+++ testing/guestbin/netbsd-prep.sh
@@ -1,12 +1,19 @@
 #!/bin/sh
 
-etc=/usr/local/etc
-rm -rf ${etc}/ipsec.*
-mkdir ${etc}/ipsec.d
-
 hostname=$(hostname)
 
+#
+# Libreswan
+#
+# install stuff into ${etc}/ipsec.*
+#
+
+etc=/usr/local/etc
+rm -rf ${etc}/ipsec.d
+mkdir ${etc}/ipsec.d
+
 for s in conf secrets ; do
+    rm -f ${etc}/ipsec.${s}
     for f in ${hostname}.${s} ipsec.${s} ; do
 	if test -r "${f}" ; then
 	    cp -v ${f} ${etc}/ipsec.${s}
@@ -14,7 +21,30 @@ for s in conf secrets ; do
     done
 done
 
-rm -f /tmp/pluto.log
-ln -s $PWD/OUTPUT/${hostname}.pluto.log /tmp/pluto.log
+if test -r ${etc}/ipsec.conf ; then
+    rm -f /tmp/pluto.log
+    ln -s $PWD/OUTPUT/${hostname}.pluto.log /tmp/pluto.log
+fi
+
+#
+# Others
+#
+# install stuff into /etc/*.conf
+#
+
+for p in iked ; do
+    rm -rf /etc/${p}.conf
+    for f in ${hostname}.${p} ${p}.conf ; do
+	if test -r "${f}" ; then
+	    cp -v ${f} /etc/${p}.conf
+	    chmod 600 /etc/${p}.conf
+	fi
+    done
+done
+
+if test -r /etc/iked.conf ; then
+    rm -f /tmp/iked.log
+    ln -s $PWD/OUTPUT/${hostname}.iked.log /tmp/iked.log
+fi
 
 stty -oxtabs
